+++ laktory/models/resources/databricks/notebook.py
@@ -54,6 +54,10 @@ class Notebook(BaseModel, PulumiResource, TerraformResource):
     path: str = None
     source: str
 
+    @classmethod
+    def lookup_id_alias(cls) -> str:
+        return "source"
+
     @property
     def filename(self) -> str:
         """Notebook file name"""