+++ laktory/models/resources/databricks/workspacefile.py
@@ -32,6 +32,10 @@ class WorkspaceFile(BaseModel, PulumiResource, TerraformResource):
     path: str = None
     source: str
 
+    @classmethod
+    def lookup_id_alias(cls) -> str:
+        return "source"
+
     @property
     def filename(self) -> str:
         """File filename"""